Fix imperfect focusing on a selected slide
When a user click a slide in overview mode, the selected slide should be zoomed
in as much as possible.  However, previous implementation considered the
enlarged page number element inside the slide element, and the bounding box of
the slide was calculated to be slightly larger, which resulted in room around
the slide after it was focused.